// TennisMatchGenerator/src/model/StatisticSourceRecord.ts
export class StatisticSourceRecord {
    number: number = 0;
    court: number = 0;
    set1: string = "";
    player1HomeId: string = "";
    player2HomeId: string = "";
    player1GuestId: string = "";
    player2GuestId: string = "";
    matchDayDate: Date | null = null;
    matchDayId: string = "";
}